# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: Url Shortener - OpenAPI 3.0
  description: API spec for the Url Shortener service.
  license:
    name: MIT
  version: 1.0.0
servers:
  - url: http://localhost:5173
    description: Local server
tags:
  - name: url
    description: Everything regarding a Url
paths:
  /api/v1/url:
    get:
      tags:
        - url
      summary: List all urls
      operationId: list_urls
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Url"
                required:
                  - data
    post:
      tags:
        - url
      summary: Creates a new short Url
      operationId: create_url
      requestBody:
        description: Create a new short url for a long url
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  $ref: "#/components/schemas/CreateUrl"
              required:
                - data
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Url"
        "400":
          description: Bad Request
        "422":
          description: Validation exception
  /api/v1/url/{id}:
    delete:
      tags:
        - url
      summary: Deletes a short Url
      operationId: delete_url
      parameters:
        - name: id
          in: path
          description: ID of short url to delete
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Successful operation
        "404":
          description: Url not found

components:
  schemas:
    Url:
      required:
        - id
        - longUrl
        - shortUrl
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        longUrl:
          type: string
          example: https://www.example.com
        shortUrl:
          type: string
          example: abc
    CreateUrl:
      description: Url object to be created
      required:
        - longUrl
      type: object
      properties:
        longUrl:
          type: string
          example: https://www.example.com